add: docker-compose file with caddy backend and frontend
What is happening here? (Architecture)
The entire frontend is built in a container and is connected as a volume
to our Caddy, which will serve our application (static server). Then, the Caddy
server itself is launched. Following this, our backend starts up with RSocket
onboard. The backend is also proxied through the Caddy server.